+++ lib/dump_viz.dart
@@ -4,7 +4,7 @@ export 'package:dump_viz/src/dependency_view.dart';
 export 'package:dump_viz/src/diff_view.dart';
 export 'package:dump_viz/src/dragdrop.dart';
 export 'package:dump_viz/src/history.dart';
-export 'package:dump_viz/src/infohelper.dart';
+export 'package:dump_viz/src/info_helper.dart';
 export 'package:dump_viz/src/process_data.dart';
 export 'package:dump_viz/src/tree_table.dart';
 export 'package:dump_viz/src/view_version.dart';
