+++ lib/src/dependency_view.dart
@@ -2,14 +2,14 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library dependency_view;
+library dump_viz.dependency_view;
 
 import 'dart:html' hide Selection;
 
 import 'package:polymer/polymer.dart';
 
 import 'history.dart';
-import 'infohelper.dart';
+import 'info_helper.dart';
 
 @CustomTag('dependency-view')
 class DependencyView extends PolymerElement {
