+++ lib/src/view_version.dart
@@ -8,7 +8,7 @@ import 'dart:html';
 
 import 'dependency_view.dart';
 import 'history.dart';
-import 'infohelper.dart';
+import 'info_helper.dart';
 import 'logical_row.dart';
 import 'tree_table.dart';
 import 'util.dart';
